codes/Dataset/dataset_GEN/requirements.txt
langchain-community
PyPDF
ragas==0.2.3
langchain_openai
pandas
PyPDF2
langchain-text-splitters
langchain-google-genai
